+++ photutils/aperture/rectangle.py
@@ -9,8 +9,9 @@ import math
 import astropy.units as u
 import numpy as np
 
-from .attributes import (ScalarAngle, PixelPositions, PositiveScalar, Scalar,
-                         SkyCoordPositions, ScalarAngleOrPixel)
+from .attributes import (ScalarAngle, PixelPositions, PositiveScalar,
+                         SkyCoordPositions, ScalarAngleOrPixel,
+                         ScalarAngleOrValue)
 from .core import PixelAperture, SkyAperture
 from .mask import ApertureMask
 from ..geometry import rectangular_overlap_grid
@@ -89,13 +90,14 @@ class RectangularMaskMixin:
             ny, nx = bbox.shape
             mask = rectangular_overlap_grid(edges[0], edges[1], edges[2],
                                             edges[3], nx, ny, w, h,
-                                            self.theta, 0, subpixels)
+                                            self._theta_radians, 0, subpixels)
 
             # subtract the inner circle for an annulus
             if hasattr(self, 'w_in'):
                 mask -= rectangular_overlap_grid(edges[0], edges[1], edges[2],
                                                  edges[3], nx, ny, self.w_in,
-                                                 self.h_in, self.theta, 0,
+                                                 self.h_in,
+                                                 self._theta_radians, 0,
                                                  subpixels)
 
             masks.append(ApertureMask(mask, bbox))
@@ -167,10 +169,11 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
         The full height of the rectangle in pixels.  For ``theta=0`` the
         height side is along the ``y`` axis.
 
-    theta : float, optional
-        The rotation angle in radians of the rectangle "width" side from
-        the positive ``x`` axis.  The rotation angle increases
-        counterclockwise.  The default is 0.
+    theta : float or `~astropy.units.Quantity`, optional
+        The rotation angle as an angular quantity
+        (`~astropy.units.Quantity` or `~astropy.coordinates.Angle`) or
+        value in radians (as a float) from the positive ``x`` axis. The
+        rotation angle increases counterclockwise.
 
     Raises
     ------
@@ -179,23 +182,27 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
 
     Examples
     --------
+    >>> from astropy.coordinates import Angle
     >>> from photutils.aperture import RectangularAperture
+
+    >>> theta = Angle(80, 'deg')
     >>> aper = RectangularAperture([10., 20.], 5., 3.)
-    >>> aper = RectangularAperture((10., 20.), 5., 3., theta=np.pi)
+    >>> aper = RectangularAperture((10., 20.), 5., 3., theta=theta)
 
     >>> pos1 = (10., 20.)  # (x, y)
     >>> pos2 = (30., 40.)
     >>> pos3 = (50., 60.)
     >>> aper = RectangularAperture([pos1, pos2, pos3], 5., 3.)
-    >>> aper = RectangularAperture((pos1, pos2, pos3), 5., 3., theta=np.pi)
+    >>> aper = RectangularAperture((pos1, pos2, pos3), 5., 3., theta=theta)
     """
 
     _params = ('positions', 'w', 'h', 'theta')
     positions = PixelPositions('The center pixel position(s).')
     w = PositiveScalar('The full width in pixels.')
     h = PositiveScalar('The full height in pixels.')
-    theta = Scalar('The counterclockwise rotation angle in radians of the '
-                   'rectangle "width" side from the positive x axis.')
+    theta = ScalarAngleOrValue('The counterclockwise rotation angle as an '
+                               'angular Quantity or value in radians from '
+                               'the positive x axis.')
 
     def __init__(self, positions, w, h, theta=0.):
         self.positions = positions
@@ -205,7 +212,7 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
 
     @property
     def _xy_extents(self):
-        return self._calc_extents(self.w, self.h, self.theta)
+        return self._calc_extents(self.w, self.h, self._theta_radians)
 
     @property
     def area(self):
@@ -237,10 +244,10 @@ class RectangularAperture(RectangularMaskMixin, PixelAperture):
         xy_positions, patch_kwargs = self._define_patch_params(origin=origin,
                                                                **kwargs)
         xy_positions = self._lower_left_positions(xy_positions, self.w,
-                                                  self.h, self.theta)
+                                                  self.h, self._theta_radians)
 
         patches = []
-        theta_deg = self.theta * 180. / np.pi
+        theta_deg = self._theta_radians * 180. / np.pi
         for xy_position in xy_positions:
             patches.append(mpatches.Rectangle(xy_position, self.w, self.h,
                                               theta_deg, **patch_kwargs))
@@ -309,10 +316,11 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
 
         For ``theta=0`` the height side is along the ``y`` axis.
 
-    theta : float, optional
-        The rotation angle in radians of the rectangle "width" side from
-        the positive ``x`` axis.  The rotation angle increases
-        counterclockwise.  The default is 0.
+    theta : float or `~astropy.units.Quantity`, optional
+        The rotation angle as an angular quantity
+        (`~astropy.units.Quantity` or `~astropy.coordinates.Angle`) or
+        value in radians (as a float) from the positive ``x`` axis. The
+        rotation angle increases counterclockwise.
 
     Raises
     ------
@@ -326,15 +334,18 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
 
     Examples
     --------
+    >>> from astropy.coordinates import Angle
     >>> from photutils.aperture import RectangularAnnulus
+
+    >>> theta = Angle(80, 'deg')
     >>> aper = RectangularAnnulus([10., 20.], 3., 8., 5.)
-    >>> aper = RectangularAnnulus((10., 20.), 3., 8., 5., theta=np.pi)
+    >>> aper = RectangularAnnulus((10., 20.), 3., 8., 5., theta=theta)
 
     >>> pos1 = (10., 20.)  # (x, y)
     >>> pos2 = (30., 40.)
     >>> pos3 = (50., 60.)
     >>> aper = RectangularAnnulus([pos1, pos2, pos3], 3., 8., 5.)
-    >>> aper = RectangularAnnulus((pos1, pos2, pos3), 3., 8., 5., theta=np.pi)
+    >>> aper = RectangularAnnulus((pos1, pos2, pos3), 3., 8., 5., theta=theta)
     """
 
     _params = ('positions', 'w_in', 'w_out', 'h_in', 'h_out', 'theta')
@@ -343,8 +354,9 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
     w_out = PositiveScalar('The outer full width in pixels.')
     h_in = PositiveScalar('The inner full height in pixels.')
     h_out = PositiveScalar('The outer full height in pixels.')
-    theta = Scalar('The counterclockwise rotation angle in radians of the '
-                   'rectangle "width" side from the positive x axis.')
+    theta = ScalarAngleOrValue('The counterclockwise rotation angle as an '
+                               'angular Quantity or value in radians from '
+                               'the positive x axis.')
 
     def __init__(self, positions, w_in, w_out, h_out, h_in=None, theta=0.):
         if not w_out > w_in:
@@ -366,7 +378,7 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
 
     @property
     def _xy_extents(self):
-        return self._calc_extents(self.w_out, self.h_out, self.theta)
+        return self._calc_extents(self.w_out, self.h_out, self._theta_radians)
 
     @property
     def area(self):
@@ -399,14 +411,14 @@ class RectangularAnnulus(RectangularMaskMixin, PixelAperture):
                                                                **kwargs)
         inner_xy_positions = self._lower_left_positions(xy_positions,
                                                         self.w_in, self.h_in,
-                                                        self.theta)
+                                                        self._theta_radians)
         outer_xy_positions = self._lower_left_positions(xy_positions,
                                                         self.w_out,
                                                         self.h_out,
-                                                        self.theta)
+                                                        self._theta_radians)
 
         patches = []
-        theta_deg = self.theta * 180. / np.pi
+        theta_deg = self._theta_radians * 180. / np.pi
         for xy_in, xy_out in zip(inner_xy_positions, outer_xy_positions):
             patch_inner = mpatches.Rectangle(xy_in, self.w_in, self.h_in,
                                              theta_deg)
@@ -467,8 +479,7 @@ class SkyRectangularAperture(SkyAperture):
     theta : scalar `~astropy.units.Quantity`, optional
         The position angle (in angular units) of the rectangle "width"
         side.  For a right-handed world coordinate system, the position
-        angle increases counterclockwise from North (PA=0).  The default
-        is 0 degrees.
+        angle increases counterclockwise from North (PA=0).
 
     Examples
     --------
@@ -557,8 +568,7 @@ class SkyRectangularAnnulus(SkyAperture):
     theta : scalar `~astropy.units.Quantity`, optional
         The position angle (in angular units) of the rectangle "width"
         side.  For a right-handed world coordinate system, the position
-        angle increases counterclockwise from North (PA=0).  The default
-        is 0 degrees.
+        angle increases counterclockwise from North (PA=0).
 
     Examples
     --------