+++ photutils/aperture/tests/test_rectangle.py
@@ -3,7 +3,7 @@
 Tests for the rectangle module.
 """
 
-from astropy.coordinates import SkyCoord
+from astropy.coordinates import Angle, SkyCoord
 import astropy.units as u
 import numpy as np
 import pytest
@@ -113,3 +113,28 @@ class TestSkyRectangularAnnulus(BaseTestAperture):
         assert aper == self.aperture
         aper.w_in = 2. * UNIT
         assert aper != self.aperture
+
+
+def test_rectangle_theta_quantity():
+    aper1 = RectangularAperture(POSITIONS, w=10., h=5., theta=np.pi/2.)
+    theta = u.Quantity(90 * u.deg)
+    aper2 = RectangularAperture(POSITIONS, w=10., h=5., theta=theta)
+    theta = Angle(90 * u.deg)
+    aper3 = RectangularAperture(POSITIONS, w=10., h=5., theta=theta)
+
+    assert aper1._theta_radians == aper2._theta_radians
+    assert aper1._theta_radians == aper3._theta_radians
+
+
+def test_rectangle_annulus_theta_quantity():
+    aper1 = RectangularAnnulus(POSITIONS, w_in=10., w_out=20., h_out=17,
+                               theta=np.pi/3)
+    theta = u.Quantity(60 * u.deg)
+    aper2 = RectangularAnnulus(POSITIONS, w_in=10., w_out=20., h_out=17,
+                               theta=theta)
+    theta = Angle(60 * u.deg)
+    aper3 = RectangularAnnulus(POSITIONS, w_in=10., w_out=20., h_out=17,
+                               theta=theta)
+
+    assert aper1._theta_radians == aper2._theta_radians
+    assert aper1._theta_radians == aper3._theta_radians